+++ src/main/CharacterSets.ts
@@ -18,6 +18,7 @@ export class CharacterSet {
   CONTRACTION_50: string;
   CONTRACTION_60: string;
   CONTRACTION_70: string;
+  CONTRACTION_80: string;
   useTraditional: boolean;
   useCapital: boolean;
 
@@ -41,14 +42,15 @@ export class CharacterSet {
     this.CONTRACTION_50 = contraction[3];
     this.CONTRACTION_60 = contraction[4];
     this.CONTRACTION_70 = contraction[5];
+    this.CONTRACTION_80 = contraction[6];
   }
 }
 
 const allCharacterSets = [
   //
-  new CharacterSet('零一二三四五六七八九', '十百千万亿兆', '正负', '点', '廿卅卌圩圆进', false, false),
+  new CharacterSet('零一二三四五六七八九', '十百千万亿兆', '正负', '点', '廿卅卌圩圆进枯', false, false),
   new CharacterSet('零壹贰叁肆伍陆柒捌玖', '拾佰仟萬億兆', '正负', '点', '念', false, true),
-  new CharacterSet('零一二三四五六七八九', '十百千萬億兆', '正負', '點', '廿卅卌圩圓進', true, false),
+  new CharacterSet('零一二三四五六七八九', '十百千萬億兆', '正負', '點', '廿卅卌圩圓進枯', true, false),
   new CharacterSet('零壹貳參肆伍陸柒捌玖', '拾佰仟萬億兆', '正負', '點', '念', true, true),
 ];