+++ src/test/data-contraction-all.ts
@@ -4,5 +4,6 @@ import {data as data_40} from './data-contraction-40';
 import {data as data_50} from './data-contraction-50';
 import {data as data_60} from './data-contraction-60';
 import {data as data_70} from './data-contraction-70';
+import {data as data_80} from './data-contraction-80';
 
-export const data = [...data_20, ...data_30, ...data_40, ...data_50, ...data_60, ...data_70];
+export const data = [...data_20, ...data_30, ...data_40, ...data_50, ...data_60, ...data_70, ...data_80];