+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - `#56`, as user, i want to optionally use contraction 50.
 - `#57`, as user, i want to optionally use contraction 60.
 - `#58`, as user, i want to optionally use contraction 70.
+- `#59`, as user, i want to optionally use contraction 80.
 
 ## `1.8.0` - 5 July 2022